my_web_app/app/requirements.txt
Flask==2.2.2
mysqlclient==2.2.0
werkzeug==2.2.2
mysql-connector-python==8.0.28  # Add this line

